requirements.txt
altair==5.1.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.23.4
tensorflow-cpu==2.14.0
streamlit==1.28.2
torch==1.12.0
torchvision==0.13.0